+++ kie-wb-common-stunner/kie-wb-common-stunner-core/kie-wb-common-stunner-commons/kie-wb-common-stunner-core-common/src/main/java/org/kie/workbench/common/stunner/core/graph/processing/layout/sugiyama/step03/LayerCrossingCount.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.workbench.common.stunner.core.graph.processing.layout.sugyiama.step03;
+package org.kie.workbench.common.stunner.core.graph.processing.layout.sugiyama.step03;
 
 import java.util.ArrayList;
 import java.util.List;
